requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
catboost==1.2.8
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
fastapi==0.116.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
gensim==4.3.3
GPUtil==1.4.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.25.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lightgbm==4.6.0
marisa-trie==1.2.1
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.3
numba==0.61.2
numpy==1.26.4
openai==1.76.0
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
pipreqs==0.5.0
plotly==6.0.1
preshed==3.0.10
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shap==0.47.1
shape==0.1.1
singlestoredb==1.12.4
six==1.17.0
sklearn-compat==0.1.3
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparams==6.2.0
srsly==2.5.1
starlette==0.47.1
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.5.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchtyping==0.1.5
torchvision==0.21.0+cu124
tqdm==4.67.1
triton==3.2.0
typer==0.16.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xgboost==3.0.0
